use u32 instead of usize in groupby and hash indexes to reduce memory usage and improve cache performance

+++ polars/polars-core/src/chunked_array/ops/mod.rs
@@ -490,7 +490,7 @@ pub trait ChunkUnique<T> {
 
     /// Get first index of the unique values in a `ChunkedArray`.
     /// This Vec is sorted.
-    fn arg_unique(&self) -> Result<Vec<usize>>;
+    fn arg_unique(&self) -> Result<Vec<u32>>;
 
     /// Number of unique values in the `ChunkedArray`
     fn n_unique(&self) -> Result<usize> {
